requirements.txt
argparse==1.2.1
backports.functools-lru-cache==1.5
cycler==0.10.0
matplotlib==2.1.2
numpy==1.14.1
pandas==0.22.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
subprocess32==3.2.7
wsgiref==0.1.2
